+++ lib/c-api/src/wasm_c_api/types/memory.rs
@@ -25,7 +25,7 @@ impl WasmMemoryType {
 }
 
 #[allow(non_camel_case_types)]
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 #[repr(transparent)]
 pub struct wasm_memorytype_t {
     pub(crate) extern_type: wasm_externtype_t,
@@ -48,7 +48,7 @@ impl wasm_memorytype_t {
     }
 }
 
-wasm_declare_vec!(memorytype);
+wasm_declare_boxed_vec!(memorytype);
 
 #[no_mangle]
 pub unsafe extern "C" fn wasm_memorytype_new(limits: &wasm_limits_t) -> Box<wasm_memorytype_t> {